requirements.txt
attr==0.3.2
beautifulsoup4==4.12.3
ConfigParser==7.0.0
contextlib2==21.6.0
cryptography==42.0.5
docutils==0.21.2
filelock==3.15.1
Flask==3.0.3
Flask_Cors==4.0.0
Flask_PyMongo==2.3.0
HTMLParser==0.0.2
HTMLParser==0.0.2
importlib_metadata==6.11.0
ipython==8.12.3
ipywidgets==8.1.3
Jinja2==3.1.2
keyring==25.2.1
langchain==0.0.316
mock==5.1.0
nltk==3.8.1
openai==0.28.1
Pillow==10.3.0
protobuf==3.20.3
pymongo==4.6.1
pyOpenSSL==24.1.0
pyOpenSSL==24.1.0
PyPDF2==3.0.1
python-dotenv==1.0.1
railroad==0.5.0
redis==5.0.6
sec_api==1.0.17
Sphinx==7.3.7
tabula_py==2.9.0
thread==2.0.3
tiktoken==0.6.0
tornado==6.4
urllib3_secure_extra==0.1.0
Werkzeug==3.0.3
PyMuPDF==1.24.2

